+++ Source/levels/themes.cpp
@@ -13,6 +13,7 @@
 #include "engine/points_in_rectangle_range.hpp"
 #include "engine/random.hpp"
 #include "items.h"
+#include "levels/tile_properties.hpp"
 #include "levels/trigs.h"
 #include "monster.h"
 #include "objects.h"
